to be continued

+++ Homework 1/OtherTasks/src/main/java/org/example/Task5/Converter.java
@@ -6,7 +6,7 @@ public class Converter {
     public static void main(String[] args) {
         System.out.println("Please enter the amount of EUR: ");
         Scanner scanner = new Scanner(System.in);
-        int amountOfEuro = scanner.nextInt();
+        double amountOfEuro = scanner.nextDouble();
         double amountOfDol = Math.round(amountOfEuro* 1.09*100.0)/100.0;
         System.out.println("According to the current euro exchange rate 1 EUR = 1.09 USD.");
         System.out.println("Your amount is equal: " + amountOfDol+ " USD");
